24.04.2020

+++ Model/Database/Receipt.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
-using System.Text;
 
 
 namespace Model.Database {
@@ -12,14 +10,14 @@ namespace Model.Database {
         public DateTime? CloseDate { get; set; }
 
         [NotMapped]
-        public float TotalPrice { get; set; }
-        public void GetTotalPrice(Context context) {
-            TotalPrice = context.Transactions.Where(e => e.Receipt == this).Sum(e => e.Service.ServicePrice);
+        public decimal TotalPrice { get; set; }
+        public void GetTotalPrice (Context context) {
+            TotalPrice = context.Transactions.Where (e => e.Receipt == this).Sum (e => e.Service.ServicePrice);
         }
 
         public int CustomerId { get; set; }
 
-        [ForeignKey(nameof(CustomerId))]
+        [ForeignKey (nameof (CustomerId))]
         virtual public Customer Customer { get; set; }
     }
 }
